pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "novel-testbed"
version = "0.2.0"
description = "A narrative compiler and test harness for fiction modules"
readme = "README.md"
requires-python = ">=3.10"

authors = [
  { name = "Matt Briggs" }
]

dependencies = [
  "pyyaml>=6.0",
  "openai>=1.3.0",
]

classifiers = [
  "Development Status :: 3 - Alpha",
  "Programming Language :: Python :: 3",
  "Programming Language :: Python :: 3.10",
  "Topic :: Software Development :: Testing",
  "Topic :: Text Processing :: Linguistic",
  "Topic :: Artistic Software",
  "Operating System :: OS Independent",
]

[project.optional-dependencies]
dev = [
  "pytest>=7.4",
  "ruff>=0.1.6",
]

docs = [
  "mkdocs>=1.5",
  "mkdocs-material>=9.5",
  "mkdocstrings[python]>=0.24",
  "mkdocs-mermaid2-plugin>=1.1",
]

uml = [
  "pylint>=3.0",
]

all = [
  "pytest>=7.4",
  "ruff>=0.1.6",
  "mkdocs>=1.5",
  "mkdocs-material>=9.5",
  "mkdocstrings[python]>=0.24",
  "mkdocs-mermaid2-plugin>=1.1",
  "pylint>=3.0",
]

[project.scripts]
novel-testbed = "novel_testbed.cli:main"